refactor(ai): rename chat/providers.ts to chat/models.ts
The file no longer exports any provider concept, only APP_MODELS and
DEFAULT_MODEL. Rename it and its four importers to match what it holds now.
