+++ apps/opensidian/src/lib/chat/models.ts
